+++ src/Api/Request/AssertRefundService.php
@@ -23,6 +23,7 @@
 
 namespace Invertus\SaferPay\Api\Request;
 
+use Exception;
 use Invertus\SaferPay\Api\ApiRequest;
 use Invertus\SaferPay\DTO\Request\Assert\AssertRequest;
 use Invertus\SaferPay\DTO\Request\AssertRefund\AssertRefundRequest;
@@ -33,7 +34,6 @@ use Invertus\SaferPay\EntityBuilder\SaferPayAssertRefundBuilder;
 use Invertus\SaferPay\Exception\Api\SaferPayApiException;
 use Invertus\SaferPay\Service\Response\AssertRefundResponseObjectCreator;
 use Invertus\SaferPay\Service\Response\AssertResponseObjectCreator;
-use PHPUnit\Exception;
 use SaferPayOrder;
 
 class AssertRefundService
@@ -44,10 +44,7 @@ class AssertRefundService
      * @var ApiRequest
      */
     private $apiRequest;
-    /**
-     * @var AssertRefundResponseObjectCreator
-     */
-    private $assertRefundResponseObjectCreator;
+
     /**
      * @var SaferPayAssertRefundBuilder
      */
@@ -66,35 +63,20 @@ class AssertRefundService
     /**
      * @param AssertRefundRequest $assertRefundRequest
      *
-     * @return AssertBody
-     * @throws \Exception
+     * @return array |null
+     * @throws Exception
      */
     public function assertRefund(AssertRefundRequest $assertRefundRequest)
     {
         $assertApi = self::ASSERT_REFUND_API_TRANSACTION;
+
         try {
-            $response = $this->apiRequest->post(
+            return $this->apiRequest->post(
                 $assertApi,
-                [
-                    'body' => json_encode($assertRefundRequest->getAsArray()),
-                ]
+                $assertRefundRequest->getAsArray()
             );
         } catch (Exception $e) {
             throw new SaferPayApiException('Assert Refund API failed', SaferPayApiException::ASSERT);
         }
-
-        return json_decode($response->getBody()->getContents());
-    }
-
-    /**
-     * @param object $responseBody
-     *
-     * @return AssertRefundBody
-     */
-    public function createObjectsFromAssertRefundResponse($responseBody)
-    {
-        $assertBody = $this->assertRefundResponseObjectCreator->createAssertRefundObject($responseBody);
-
-        return $assertBody;
     }
 }