+++ src/Api/Request/ObtainPaymentMethodsService.php
@@ -23,6 +23,7 @@
 
 namespace Invertus\SaferPay\Api\Request;
 
+use Exception;
 use Invertus\SaferPay\Api\ApiRequest;
 use Invertus\SaferPay\DTO\Request\ObtainPaymentMethods\ObtainPaymentMethodsRequest;
 
@@ -36,10 +37,11 @@ class ObtainPaymentMethodsService
         $this->apiRequest = $apiRequest;
     }
 
+    /**
+     * @throws Exception
+     */
     public function getPaymentMethods(ObtainPaymentMethodsRequest $obtainPaymentMethodsRequestDTO)
     {
-        $response = $this->apiRequest->get($obtainPaymentMethodsRequestDTO->generateRequestUrl());
-
-        return json_decode($response->getBody()->getContents());
+        return $this->apiRequest->get($obtainPaymentMethodsRequestDTO->generateRequestUrl());
     }
 }