+++ src/DTO/Request/ObtainPaymentMethods/ObtainPaymentMethodsRequest.php
@@ -43,6 +43,6 @@ class ObtainPaymentMethodsRequest
     {
         $customerId = $this->requestHeader->getCustomerId();
 
-        return "/api/rest/customers/{$customerId}/terminals/{$this->terminalId}/payment-methods";
+        return "rest/customers/{$customerId}/terminals/{$this->terminalId}/payment-methods";
     }
 }