+++ src/Factory/ModuleFactory.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Invertus\SaferPay\Factory;
+
+use Module;
+
+class ModuleFactory
+{
+    /**
+     * @return \SaferPayOfficial|Module|null
+     */
+    public function getModule()
+    {
+        return Module::getInstanceByName('saferpayofficial');
+    }
+}