+++ src/Service/SaferPayExceptionService.php
@@ -25,6 +25,7 @@ namespace Invertus\SaferPay\Service;
 
 use Exception;
 use Invertus\SaferPay\Exception\Api\SaferPayApiException;
+use Invertus\SaferPay\Factory\ModuleFactory;
 use SaferPayOfficial;
 
 class SaferPayExceptionService
@@ -36,9 +37,9 @@ class SaferPayExceptionService
      */
     private $module;
 
-    public function __construct(SaferPayOfficial $module)
+    public function __construct(ModuleFactory $module)
     {
-        $this->module = $module;
+        $this->module = $module->getModule();
     }
 
     public function getErrorMessages()